Update the SessionHelper and Sleep Time for Actions

+++ src/common/UITestAutomation/SessionHelper.cs
@@ -31,9 +31,12 @@ namespace Microsoft.PowerToys.UITest
         private Process? appDriver;
         private Process? runner;
 
+        private PowerToysModule scope;
+
         [UnconditionalSuppressMessage("SingleFile", "IL3000:Avoid accessing Assembly file path when publishing as a single file", Justification = "<Pending>")]
         public SessionHelper(PowerToysModule scope)
         {
+            this.scope = scope;
             this.sessionPath = ModuleConfigData.Instance.GetModulePath(scope);
             this.locationPath = Path.GetDirectoryName(Assembly.GetExecutingAssembly().Location);
 
@@ -52,7 +55,10 @@ namespace Microsoft.PowerToys.UITest
                 Verb = "runas",
             };
 
-            this.runner = Process.Start(runnerProcessInfo);
+            if (scope == PowerToysModule.PowerToysSettings)
+            {
+                this.runner = Process.Start(runnerProcessInfo);
+            }
 
             var desktopCapabilities = new AppiumOptions();
             desktopCapabilities.AddAdditionalCapability("app", "Root");
@@ -83,8 +89,11 @@ namespace Microsoft.PowerToys.UITest
             {
                 appDriver?.Kill();
                 appDriver?.WaitForExit(); // Optional: Wait for the process to exit
-                runner?.Kill();
-                runner?.WaitForExit(); // Optional: Wait for the process to exit
+                if (this.scope == PowerToysModule.PowerToysSettings)
+                {
+                    runner?.Kill();
+                    runner?.WaitForExit(); // Optional: Wait for the process to exit
+                }
             }
             catch (Exception ex)
             {
